requirements.txt
fastapi>=0.128,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.10,<3.0
python-dotenv>=1.0,<2.0
google-genai>=1.5,<2.0
anthropic>=0.42,<1.0
pillow>=11.0,<13.0
markdown>=3.7,<4.0
